requirements.txt
python==3.8.10
discord.py==2.0.1
git+https://github.com/Rapptz/discord-ext-menus
python-dotenv==0.21.0
roblox=2.0.0
chat-exporter